Fix compilation warnings on 64 bit architectures
The ErlDrvEvent is a pointer, holding socket file descriptors, which
are ints. On 64 bit machines, the size of a pointer is larger than the
size of an int, so an direct cast like ev = (ErlDrvEvent)some_int
results in the following warning:

   warning: cast to pointer from integer of different size

This commit eliminates that warning on both 32 and 64 bit machines.